+++ package.json
@@ -14,7 +14,6 @@
   "dependencies": {
     "@azure/storage-blob": "^12.25.0",
     "@hapi/boom": "^10.0.1",
-    "busboy": "^1.6.0",
     "cookie-parser": "^1.4.6",
     "curp": "^1.2.3",
     "dotenv": "^16.4.5",
@@ -24,6 +23,7 @@
     "jsonwebtoken": "^9.0.2",
     "jwt-decode": "^4.0.0",
     "morgan": "^1.10.0",
+    "multer": "^1.4.5-lts.1",
     "mysql2": "^3.11.3",
     "passport": "^0.7.0",
     "passport-azure-ad": "^4.3.5",
