+++ src/middlewares/index.js
@@ -3,7 +3,7 @@ export * from './checkRoleAdmin.js';
 export * from './cors.js';
 export * from './decodeToken.js';
 export * from './errorHandler.js';
-export * from './formDataParser.js';
 export * from './msalTokenVerify.js';
+export * from './multer.js';
 export * from './notTheSameUser.js';
 export * from './userAllowedHandler.js';
