+++ src/routes/students.router.js
@@ -1,4 +1,5 @@
 import { Router } from "express";
+import { uploadDocs } from "../middlewares/index.js";
 import {
   dataGeneral,
   dbStudent,
@@ -6,13 +7,12 @@ import {
   inscription,
   typeRegister,
 } from "../controller/students.controller.js";
-import { formDataParser } from "../middlewares/index.js";
 
 const router = Router();
 router.get("/typeRegister/:curp", typeRegister);
-router.post("/newStudent/dataGeneral", formDataParser, dataGeneral);
-router.post("/newStudent/inscription", formDataParser, inscription);
-router.post("/DBStudent", formDataParser, dbStudent);
-router.post("/files", formDataParser, files);
+router.post("/newStudent/dataGeneral", uploadDocs, dataGeneral);
+router.post("/newStudent/inscription", uploadDocs, inscription);
+router.post("/DBStudent", uploadDocs, dbStudent);
+router.post("/files", uploadDocs, files);
 
 export default router;
